pyproject.toml
[tool.poetry]
name = "v1-162-part12"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.30.1"
langserve = "^0.3.0"
pydantic = "^2.7.4"
extraction-openai-functions = {path = "packages/extraction-openai-functions", develop = true}
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
antlr4-python3-runtime = "4.9.3"
anyio = "4.3.0"
attrs = "23.2.0"
backoff = "2.2.1"
beautifulsoup4 = "4.12.3"
certifi = "2024.2.2"
cffi = "1.16.0"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
coloredlogs = "15.0.1"
contourpy = "1.2.0"
cryptography = "42.0.2"
cycler = "0.12.1"
dataclasses-json = "0.6.4"
dataclasses-json-speakeasy = "0.5.11"
deprecated = "1.2.14"
distro = "1.9.0"
effdet = "0.4.1"
emoji = "2.10.1"
et-xmlfile = "1.1.0"
fastapi = "^0.111.0"
filelock = "3.13.1"
filetype = "1.2.0"
flatbuffers = "23.5.26"
fonttools = "4.49.0"
frozenlist = "1.4.1"
fsspec = "2024.2.0"
gitdb = "4.0.11"
gitpython = "3.1.42"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
httpx-sse = "0.4.0"
huggingface-hub = "0.20.3"
humanfriendly = "10.0"
idna = "3.6"
importlib-metadata = "7.0.1"
importlib-resources = "6.1.1"
iopath = "0.1.10"
jinja2 = "3.1.3"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpath-python = "1.0.6"
jsonpointer = "2.4"
kiwisolver = "1.4.5"
langchain = "^0.3.0"
#langchain-cli = "0.0.21"
langchain-text-splitters = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.3.0"
langdetect = "1.0.9"
langsmith = "^0.3.0"
layoutparser = "0.3.4"
lxml = "5.1.0"
markdown = "3.5.2"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.20.2"
matplotlib = "3.7.2"
mdurl = "0.1.2"
mpmath = "1.3.0"
msg-parser = "1.2.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
networkx = "3.2.1"
nltk = "3.8.1"
numpy = "1.26.4"
olefile = "0.47"
omegaconf = "2.3.0"
onnx = "1.15.0"
onnxruntime = "1.18.0"
openai = "^1.86.0"
opencv-python = "4.8.0.76"
openpyxl = "3.1.2"
orjson = "3.10.0"
packaging = "23.2"
pandas = "2.2.0"
pdf2image = "1.17.0"
pdfminer-six = "^20231228"
pdfplumber = "^0.11.0"
pgvector = "0.2.5" # Keep this version for langchain_community.vectorstores.pgvector
pikepdf = "8.11.0"
pillow = "10.2.0"
pillow-heif = "0.15.0"
pi-heif = "^0.16.0"
portalocker = "2.8.2"
protobuf = "4.23.4"
psycopg = {extras = ["binary", "async"], version = "^3.2.9"} # Use psycopg for both sync and async
pycocotools = "2.0.7"
pycparser = "2.21"
pygments = "2.17.2"
pypandoc = "1.12"
pyparsing = "3.0.9"
pypdf = "4.0.1"
pypdfium2 = "4.27.0"
pytesseract = "0.3.10"
python-dateutil = "2.8.2"
python-docx = "1.1.0"
python-dotenv = "1.0.1"
python-iso639 = "2024.2.7"
python-magic = "0.4.27"
python-multipart = "0.0.9"
python-pptx = "0.6.23"
pytz = "2024.1"
pyyaml = "6.0.1"
rapidfuzz = "3.6.1"
regex = "2023.12.25"
requests = "2.31.0"
rich = "13.7.1"
safetensors = "0.3.2"
scipy = "1.10.1"
shellingham = "1.5.4"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
soupsieve = "2.5"
sqlalchemy = "2.0.29"
sse-starlette = "1.8.2"
starlette = "^0.37.2"
sympy = "1.12"
tabulate = "0.9.0"
tenacity = "8.2.3"
tiktoken = "^0.7"
timm = "0.9.12"
tokenizers = "0.15.2"
tomlkit = "0.12.4"
torch = "2.2.0"
torchvision = "0.17.0"
tqdm = "4.66.2"
transformers = "4.37.1"
typer = "^0.15.1"
typing-inspect = "0.9.0"
typing-extensions = "^4.11"
tzdata = "2024.1"
unstructured = "^0.16.0"
unstructured-client = "0.18.0"
unstructured-inference = "0.8.10"
unstructured-pytesseract = "0.3.12"
urllib3 = "1.26.18"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websockets = "12.0"
wrapt = "1.16.0"
xlrd = "2.0.1"
xlsxwriter = "3.1.9"
yarl = "1.9.4"
zipp = "3.17.0"
greenlet = "^1.0.0"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"